Add function __zip_file() to read zip files
Because of the nature of ZipFile returning IO[bytes] when opening a member of a zip file, the typing hints for the other functions needed to be updated.

Additionally, since __read_file() doesn't need to be called more than once, the main logic of deciding which function to use to load a file got put into __load_file().

Also, the test_*() functions don't actually test for correct configuration usage, so that needs to be fixed.